refactor(helpers): rename helpers.formatting to helpers.labels
src/dartwork_mpl/ had two different modules sharing the name
'formatting':

  - dartwork_mpl.formatting          (7 axis tick formatters;
                                       re-exported as dm.format_axis_*)
  - dartwork_mpl.helpers.formatting  (3 label / legend / value-label
                                       composition helpers)

Different namespaces so no runtime clash, but a foot-gun when
navigating the source and when greppping for 'formatting'. The
helpers version additionally did `import dartwork_mpl as dm` from
inside a subpackage of dartwork_mpl — a self-import that only worked
because the references were inside function bodies.

Rename the subpackage module to `labels`, which is a fair descriptor
for its three functions (format_axis_labels, optimize_legend,
add_value_labels). Keep the top-level `dartwork_mpl.formatting`
untouched (its `dm.format_axis_*` contract is documented and widely
used).

- Replace `import dartwork_mpl as dm` in the renamed module with
  `from ..scale import fs`. No more self-import.
- Leave a deprecation shim at the old path:
  `dartwork_mpl/helpers/formatting.py` now re-exports from `.labels`
  and emits `DeprecationWarning` on import.
- Update `helpers/__init__.py` to import from `labels`.
- Update `docs/api/helpers.rst` and the gallery script (renamed to
  `plot_helpers_labels.py` for consistency) to use the new path.
- The `valuation` workspace sibling does not use any of these names,
  so nothing downstream breaks.

Verified:
- `from dartwork_mpl.helpers.labels import ...` works (canonical).
- `from dartwork_mpl.helpers.formatting import ...` still works and
  emits DeprecationWarning pointing at .labels.
- `dm.helpers.labels.<fn>` and `dm.helpers.formatting.<fn>` both
  reachable.
- 532 tests pass; ruff / format / mypy all green.
- Gallery examples run under the Agg backend.

Closes #46
