+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 - **Guardrail scope expanded to `docs/` and `examples/`**: `tests/test_domain_neutrality.py` now runs a second parametrised test `test_no_finance_terms_in_docs_and_examples` that scans every `.py` / `.md` / `.rst` / `.mplstyle` file under `docs/` and the top-level `examples/` tree for the same finance-domain vocabulary (English: revenue, profit, ebitda, earnings, fiscal, valuation, dcf; Korean: 매출, 매출액, 영업이익, 억원, 조원). Build artefacts (`_build`, `examples_gallery`, `_static`, `_templates`, `__pycache__`) are excluded. The original `test_no_finance_terms_in_shipped_source` coverage is unchanged. The full guardrail now scans 184+ files per run.
+- **`dm.helpers.labels`** — new submodule name for what used to be `dm.helpers.formatting`. Houses `format_axis_labels`, `optimize_legend`, and `add_value_labels`. Renamed to avoid the naming clash with the top-level `dartwork_mpl.formatting` module (which houses the `format_axis_*` tick formatters). The old `dm.helpers.formatting` import path still works but emits a `DeprecationWarning` pointing at the new name.
 - **Gallery: six new single-plot spine / grid examples** under `docs/examples_source/01_styling_and_themes/` — `plot_spine_minimal.py`, `plot_spine_visibility.py`, `plot_spine_styling.py`, `plot_grid_customization.py`, `plot_spine_publication_styles.py`, `plot_spine_dark_theme.py`, `plot_spine_dashboard.py`. Each renders exactly one figure, per the repo's one-plot-per-file convention.
-- **Gallery: six new single-plot helpers examples** under `docs/examples_source/05_advanced_components/` — `plot_helpers_data_validation.py`, `plot_helpers_color_selection.py`, `plot_helpers_formatting.py`, `plot_helpers_quality.py`, `plot_helpers_io.py`, `plot_helpers_workflow.py`. Each renders exactly one figure and covers one `dm.helpers` submodule (plus an end-to-end workflow demo), replacing the monolithic six-figure `plot_helpers_usage.py`.
+- **Gallery: six new single-plot helpers examples** under `docs/examples_source/05_advanced_components/` — `plot_helpers_data_validation.py`, `plot_helpers_color_selection.py`, `plot_helpers_labels.py` (renamed from `plot_helpers_formatting.py` for consistency with the new `dm.helpers.labels` submodule), `plot_helpers_quality.py`, `plot_helpers_io.py`, `plot_helpers_workflow.py`. Each renders exactly one figure and covers one `dm.helpers` submodule (plus an end-to-end workflow demo), replacing the monolithic six-figure `plot_helpers_usage.py`.
 
 ### Changed
 
@@ -26,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - **`validate_enhanced` is a first-class public surface now.** `src/dartwork_mpl/__init__.py` now imports the `validate_enhanced` submodule explicitly and re-exports `validate_with_fixes` at the top level, so `dm.validate_enhanced.get_fix_suggestions` / `check_agent_requirements` / `generate_validation_report` and `dm.validate_with_fixes` all work without depending on import side effects. Both `validate_enhanced` and `validate_with_fixes` are now in `__all__`, matching what `docs/api/validate.rst` already promises. `get_fix_suggestions` / `check_agent_requirements` / `generate_validation_report` remain reachable via `dm.validate_enhanced.<name>` only — top-level surface stays narrow.
 - **`docs/usage_guide/tutorials.md` "Business Report" tutorial** is now "Korean Operations Report". The `report-kr` preset showcase previously branded its example as a quarterly revenue / margin report (`매출액 (억원)`, `영업이익률 (%)`, `quarters = ["1Q24", ...]`). Rewritten to a weekly operations report (`처리량 (건)`, `가동률 (%)`, `weeks = ["1주차", ...]`) so the pedagogical payload — Korean font + dual-axis bar+line + `PercentFormatter` + `auto_layout` — stays identical but the surrounding narrative is domain-neutral. Grid-card title, in-page anchor (`korean-operations-report`), and save filename updated accordingly.
 - **`docs/api/formatting.rst` example and section framing** re-themed to remove the implicit "formatters are a finance tool" branding. The "Financial Data Formatting" section becomes "Large Number Formatting" (sample counts / dataset rows / population instead of revenue / market cap / GDP). The "Complete Example: Financial Dashboard" becomes "Complete Example: Multi-format Dashboard" — a 2×2 grid exercising all four numeric formatters on neutral metrics (sample count, efficiency, unit price, cumulative energy). The "Financial Reports" common-patterns subsection, the "Financial formats" feature bullet, the `data_type == 'financial'` branch, and the "technical/financial" Best Practice line are all neutralised. API signatures are unchanged.
+
+### Deprecated
+
+- **`dartwork_mpl.helpers.formatting`** — deprecated alias for `dartwork_mpl.helpers.labels`. Importing from the old path still works but emits a `DeprecationWarning` pointing at the new submodule. Renamed to resolve the long-standing name clash with the top-level `dartwork_mpl.formatting` module (which houses the `format_axis_*` tick formatters). Plan to remove the deprecated shim in a future major release.
 
 ### Removed
